Fix install script version override
The install script (suggested to be used in CI pipelines) supports
version overrides. This is especially useful in environments where tools
and dependencies need to be pinned to a certain version. It seems this
has been accidentally changed by an automated CI process during the
latest release; version overrides no longer work, and instead, only the
most recent release (`v1.6.14`) can be used.

This change fixes the accidental commit by using the provided program
argument instead of the hardcoded value when an overridden version is
provided to `actionlint`.

Closes #161

Signed-off-by: Matei David <matei@buoyant.io>