ci: Fix step to update version in install script
When releasing, it seems that the GitHub Action step responsible for
updating the latest version in the download (/install) script will
replace all occurences of `version=` with the new release version.

The step has a slight hint of undefined behaviour whereby the `sed`
command that replaces the pinned version will end up replacing the
version override variable too. This means that every release,
`download-actionlint` will be broken and will not work with version
overrides. This can have consequences to environments where operators
rely on pinned versions of `actionlint`.

The introduced change alters the `sed` command in the release workflow
to only change the first occurence of the searchterm.

Signed-off-by: Matei David <matei@buoyant.io>