use `sort.Stable` instead of `sort.Sort` to ensure the order of errors is deterministic
this change is necessary due to the recent change in Go. `sort.Sort` is
now using pdqsort, which shuffles a given slice at first to defeat some
bad patterns. This means that if some errors have the same line number
and column number and file path, the order is no longer deterministic.
Stable sort with `sort.Stable` guarantees the order.