feat: Support YAML anchors and aliases in parser
This change adds support for YAML anchors and aliases in the workflow parser.

A new helper function `resolveAlias` is introduced to dereference alias nodes before they are processed by the parsing functions. This function is called at the beginning of each parsing function to ensure that the parser is always working with the actual value nodes.

A new test case is added to verify that YAML anchors are correctly parsed. An existing test is also updated to ignore the `version` field in the SARIF output to make the tests more stable.